Update date format of Get-PCUsage sample command
Previous comand sample contains the datetime "31-12-1999 00:00:00". This month number is 31. If we use this sample, we should get error about datetime format. So I modify the sample. We should also modify the following 2 error messages.

"Start time is not in a valid format. Use '31-12-1999 00:00:00' format"
"End time is not in a valid format. Use '31-12-1999 00:00:00' format"